+++ bridge/opencensus/internal/ocmetric/metric_test.go
@@ -56,7 +56,6 @@ func TestConvertMetrics(t *testing.T) {
 					},
 					TimeSeries: []*ocmetricdata.TimeSeries{
 						{
-
 							LabelValues: []ocmetricdata.LabelValue{
 								{
 									Value:   "hello",
@@ -370,7 +369,8 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-		}, {
+		},
+		{
 			desc: "histogram without data points",
 			input: []*ocmetricdata.Metric{
 				{
@@ -393,7 +393,8 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-		}, {
+		},
+		{
 			desc: "sum without data points",
 			input: []*ocmetricdata.Metric{
 				{
@@ -417,7 +418,8 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-		}, {
+		},
+		{
 			desc: "gauge without data points",
 			input: []*ocmetricdata.Metric{
 				{
@@ -439,7 +441,8 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-		}, {
+		},
+		{
 			desc: "histogram with negative count",
 			input: []*ocmetricdata.Metric{
 				{
@@ -461,8 +464,9 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-			expectedErr: errConversion,
-		}, {
+			expectedErr: errNegativeDistributionCount,
+		},
+		{
 			desc: "histogram with negative bucket count",
 			input: []*ocmetricdata.Metric{
 				{
@@ -488,8 +492,9 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-			expectedErr: errConversion,
-		}, {
+			expectedErr: errNegativeBucketCount,
+		},
+		{
 			desc: "histogram with non-histogram datapoint type",
 			input: []*ocmetricdata.Metric{
 				{
@@ -509,8 +514,9 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-			expectedErr: errConversion,
-		}, {
+			expectedErr: errMismatchedValueTypes,
+		},
+		{
 			desc: "sum with non-sum datapoint type",
 			input: []*ocmetricdata.Metric{
 				{
@@ -530,8 +536,9 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-			expectedErr: errConversion,
-		}, {
+			expectedErr: errMismatchedValueTypes,
+		},
+		{
 			desc: "gauge with non-gauge datapoint type",
 			input: []*ocmetricdata.Metric{
 				{
@@ -551,8 +558,9 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-			expectedErr: errConversion,
-		}, {
+			expectedErr: errMismatchedValueTypes,
+		},
+		{
 			desc: "unsupported Gauge Distribution type",
 			input: []*ocmetricdata.Metric{
 				{
@@ -564,7 +572,7 @@ func TestConvertMetrics(t *testing.T) {
 					},
 				},
 			},
-			expectedErr: errConversion,
+			expectedErr: errAggregationType,
 		},
 	} {
 		t.Run(tc.desc, func(t *testing.T) {