+++ exporters/otlp/otlpmetric/otlpmetricgrpc/internal/otest/client.go
@@ -38,7 +38,7 @@ import (
 
 var (
 	// Sat Jan 01 2000 00:00:00 GMT+0000.
-	start = time.Date(2000, time.January, 01, 0, 0, 0, 0, time.FixedZone("GMT", 0))
+	start = time.Date(2000, time.January, 0o1, 0, 0, 0, 0, time.FixedZone("GMT", 0))
 	end   = start.Add(30 * time.Second)
 
 	kvAlice = &cpb.KeyValue{Key: "user", Value: &cpb.AnyValue{